lambdas/requirements.txt
beautifulsoup4
requests
dnspython